bezier attribute now on transmission object
